+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 - Add field setting to restrict which column types editors can choose ([#53](https://github.com/verbb/tablemaker/issues/53)).
 - Add field settings for min/max rows and min/max columns ([#38](https://github.com/verbb/tablemaker/issues/38)).
 - Confirm before deleting a column in the Edit columns dialog ([#58](https://github.com/verbb/tablemaker/issues/58)).
-- Allow passing an attributes array (or class string) to `.table` HTML output ([#4](https://github.com/verbb/tablemaker/issues/4)).
+- Allow passing an attributes array to `.table` HTML output ([#4](https://github.com/verbb/tablemaker/issues/4)).
 
 ### Removed
 - Removed Column Label field setting.